requirements.txt
python-dotenv~=1.0.1
whitenoise
requests~=2.32.3
requests-oauthlib
cryptography

daphne
channels~=4.2.0
Django~=5.1.3
djangorestframework~=3.15.2
django-storages
drf-yasg
django-cors-headers
django-debug-toolbar

psycopg
psycopg[binary]
psycopg2-binary
boto3~=1.35.64
pandas~=2.2.3
openpyxl
xlsxwriter
xlrd

tqdm~=4.67.0
python-dateutil~=2.9.0.post0
botocore~=1.35.64
docker~=7.1.0

pytz